[ADD] Machine learning model & details
Added requirements to the requirements.txt file. Added seeds to fill data in the database synthetically. Added a machine learning model with MinMaxScaler to get similar recommendations to interests and favorite books for the users, also added a command to train the model with python manage.py train_recommendation, which generates the .pkl file with the data for the current users.